updated system prompt and schemas

+++ schemas.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, Field
 
 class EvaluationSchema(BaseModel):
-    content_score: int = Field(..., ge=1, le=5)
+    content_score: float = Field(..., ge=1.0, le=5.0)
     content_explanation: str
-    format_score: int = Field(..., ge=1, le=5)
+    format_score: float = Field(..., ge=1.0, le=5.0)
     format_explanation: str
